+++ src/services/windowController.ts
@@ -1,0 +1,5 @@
+import WindowController from "../classes/window";
+
+const windowController = new WindowController();
+
+export default windowController;