+++ src/utils/modifiers.ts
@@ -1,0 +1,1 @@
+export const heightModifier = 0.5138888888888889;